[Silver III] Title: 주유소, Time: 432 ms, Memory: 35468 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.java
@@ -18,14 +18,14 @@ public class Main {
             price[i] = Integer.parseInt(st.nextToken());
         }
 
-        int sum = 0;
+        long sum = 0;
         int i = -1;
         int minPrice = Integer.MAX_VALUE;
         while (++i < n - 1) {
             if (minPrice > price[i]) {
                 minPrice = price[i];
             }
-            sum += minPrice * dist[i];
+            sum += (long) minPrice * dist[i];
         }
 
         System.out.println(sum);
